requirements.txt
PyYAML==6.0.1
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.10.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
yacs==0.1.8
ftfy==6.1.1
regex==2023.8.8
gdown==4.7.1
